+++ history.md
@@ -1,6 +1,13 @@
 ### History
 
-2.2.0
+
+v2.3.0
+- Remove pdf and jpeg from the list of supported export formats.
+- Add Aspect ratio for plot export.
+- Add a tab for fit model style in settings.
+- Migrate from Matplotlib to PyQtGraph for time series plotting.
+
+v2.2.0
 - Improve grd file handling.
 - Update reference paper details.
 - Centralize version and date in project init.